A few minor revisions to plant ovary related terms.  Added ‘plant’ to names for consistency
Was ovary placenta (PO:0020001):
A portion of placenta tissue that is part of an ovary and to which ovules are attached.

Revised:
plant ovary placenta (PO:0020001):
A portion of placenta tissue (PO:0025078) which is part of a plant ovary (PO:0009072), and to which plant ovary ovules (PO:0025490) are attached.
PO_GIT:144

Was:
ovary replum (PO:0025272)
A portion of ovary placenta tissue that divides an ovary into two or more chambers and develops from contact of the placental tissue from opposite sides of the ovary early in gynoecium development.

Revised:
plant ovary replum (PO:0025272):
A portion of ovary placenta tissue (PO:0020001) that divides a plant ovary (PO:0009072) into two or more chambers, and develops from contact of the placental tissue from opposite sides of the plant ovary early in gynoecium development (GO:0048467).

Was:
ovary septum PO:0025262
Revised:
plant ovary septum PO:0025262:

Was:
ovary epidermis PO:0006063
Revised: plant ovary epidermis PO:0006063