Revsions to leaf sheath (PO:0020104) Current def'n: A tubular portion of the leaf surrounding the stem, as in the Poaceae.
is_a cardinal organ part (PO:0025001)
part_of  part_of some 'vascular leaf' (PO:0009025)

Revised def'n: A cardinal organ part (PO:0025001) that is the lower part of a vascular leaf (PO:0009025), and which more or less encases the stem (PO:0009047).

Current comment: If you are annotating to this term, please add an additional annotation to vascular leaf (PO:0009025) or non-vascular leaf (PO:0025075), depending on the species. All annotations for angiosperms, gymnosperms, and pteridophytes should go to vascular leaf and all annotations for bryophytes should go to non-vascular leaf.

Revised comment: A leaf sheath is usually tubular shaped, and is found in the Poaceae family and other monocots.

Removed the annotation statement as it is not necessary since this term is part of vascular leaf' (PO:0009025)