Revisions to margin (and related) terms in order to add new term for leaf sheath margin:
* Renamed: organ margin  PO:0025005 --> plant organ margin PO:0025005
revised def'n: added plant to organ for consistency

Revised def'ns of child terms  for consistency:
* leaf margin (PO:0020128)- revised EA as is was causing circularity
'plant organ margin' and (part_of some leaf)  -->  'phyllome margin' and (part_of some leaf)

* Renamed organ lamina (PO:0025513)  plant organ lamina (PO:0025513)
Revised def'n: phyllome lamina (PO:0025396): A plant organ lamina (PO:0025513) that is part of a phyllome (PO:0006001).

plant organ margin (PO:0025005): A cardinal organ part that is the edge of an organ and the area immediately internally adjacent to it.
Revised def'n to match relations: A cardinal organ part (PO:0025001) that is the edge of a plant organ (PO:0009008) and the area immediately internally adjacent to it.

* phyllome margin (PO:0025018):  A plant organ margin (PO:0025005) that is part of a phyllome (PO:0006001).
Revised:  A plant organ margin (PO:0025005) that is part of a phyllome lamina (PO:0025396).

Revised children of phyllome margin (PO:0025018)- these should really all be lamina margins.....
PO:0025011 bract margin
PO:0025008 petal margin
PO:0025013 prophyll margin
PO:0025014 scale leaf margin
PO:0005021 sepal margin
PO:0025016 sporophyll margin
PO:0025015 tepal margin

* moved cotyledon margin ( PO:0025012 ) to be a leaf margin (PO:0020128)
Revised: cotyledon margin (PO:0025012):  A leaf margin (PO:0020128) that is part of a cotyledon (PO:0020030).

* leaf margin (PO:0020128): A phyllome margin (PO:0025018) of any flattened part of a leaf (PO:0025034).
Revised:  A phyllome margin (PO:0025018) that is part of a leaf (PO:0025034).

child terms: both these are cardinal organ parts, not leaves.

* leaflet margin ( PO:0006034):  Def'n: The margin of the leaflet.
revised defn: A cardinal organ part (PO:0025001) that is the edge of a leaflet (PO:0020049) and the area immediately internally adjacent to it.

revised defn:
leaflet (PO:0020049): One of the ultimate segments of a compound leaf.
revised defn:  A cardinal organ part (PO:0025001) that is one of the ultimate segments of a compound leaf (PO:0020043).

* petiole margin (PO:0025010): .
A cardinal organ part (PO:0025001) that is the edge of a petiole lamina (PO:0025129) and the area immediately internally adjacent to it.

New term request:
leaf sheath margin (PO:0025619)

Proposed definiton: leaf sheath margin (PO:0025619): A cardinal organ part (PO:0025001) that is the edge of a leaf sheath (PO:0020104) and the area immediately internally adjacent to it.